Extract GitGitGadget's GitHub App into its own repository
To make it eminently more deployable, let's follow the example of
https://github.com/dscho/leave-me-alone-github-app how to set up a
GitHub App that is automatically deployed upon updating `main`, and
https://github.com/git-for-windows/gfw-helper-github-app as an example
how to do proper CI builds and stuff.

This is step number one: start with GitGitGadget's code, but remove all
but the Azure Function's code.

Over the course of the next commits, the GitHub workflows and the tests
will be added.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>