Add support functions to authenticate as GitHub App
In particular, this adds a function to obtain the installation access
token for a given repository. This function will be used e.g. to trigger
GitHub workflow runs.

These functions are slightly edited versions of functions that have been
in use in https://github.com/git-for-windows/gfw-helper-github-app for
quite a while. But the test is new.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>