Add a function to trigger a GitHub workflow
The GitHub workflow must have a `workflow_dispatch` trigger for this to
work. The `workflow_id` is the file name of the workflow, e.g.
`sync-ref.yml`.

This function is a slightly edited version of a function that has been
in use in https://github.com/git-for-windows/gfw-helper-github-app for
quite a while. But the test is new.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>