Move to a new org for GitGitGadget's manually-triggered workflows
The Git maintainer frequently pushes out dozens of branches at the same
time. Every of those branches triggers their own CI run, spawning almost
50 jobs, with an overall runtime of over 8h (!) which would amortize to
a shorter wall-clock time if only the concurrency limit of 20 parallel
jobs was not depleted so efficiently.

These concurrency limits are org-wide on GitHub. That is, when dozens of
workflow runs are partially running, partially queued up, in
`gitgitgadget/git`, all new workflow runs in
`gitgitgadget/gitgitgadget-workflows` are queued and will have to wait
their turn.

In combination, this leads to a frequent starvation of GitHub Actions
resources, which means: even short tasks (like the `sync-ref` workflow)
won't run for several hours in such situations.

To remedy that, I forked the `gitgitgadget-workflows` repository into a
new org: `gitgitgadget-workflows. That way, the workflows like
`sync-ref` do not have to wait for CI/PR runs to finish in
`gitgitgadget/git`.

This PR adjusts GitGitGadget's GitHub App to trigger the `sync-ref`
workflow in that org.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>