Trigger `update-prs` and `update-mail-to-commit-notes` on `seen` updates
This is part of my ongoing effort to replace GitGitGadget's Azure
Pipelines by GitHub workflows.

Currently, the Azure Pipelines have the upstream Git repository as
"source" (even if they don't check that out), i.e. Azure Pipelines poll
for changes.

In GitHub workflows, such a thing is no longer possible: There is no way
to trigger a GitHub workflow on updates in _another_ repository than the
workflow definition resides (a rather vexing limitation of GitHub
Actions' architecture, to be sure).

In this instance, we can avoid resorting to the ugly, ugly, yet
all-too-common workaround of polling (which we do in
`sync-git-mailing-list-mirror`) but can at least react to the respective
webhook event (at least we _can_, what with the `gitgitgadget-git` app
being installed on `git/git`).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>