deploy: allow deploying in forks
It would be nice if we could deploy the Azure Function contingent on the
presence of the `AZURE_CLIENT_ID` secret. However, this is not possible
in GitHub workflows: the job-level `if:` conditions lack access to the
`secrets` context. Strangely enough, they _do_ have access to the `vars`
context...

To successfully deploy the Azure Function, it needs to know which
`gitgitgadget-workflows` fork to target when triggering workflow runs,
anyway, so let's _require_ a repository variable called
`DEPLOY_WITH_WORKFLOWS` that specifies that fork in the form
`<org>/gitgitgadget-workflows`.

Note that such a `gitgitgadget-workflows` fork _must_ have the `config`
branch, with a `gitgitgadget-config.json` file that contains the
corresponding project configuration; The `deploy` workflow will retrieve
this configuration and overwrite `gitgitgadget-config.json` with it,
augmenting the `workflowsRepo` information on the fly.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>