Azure Function: avoid errors about `context.done()`
An Azure Function is supposed to either return a `Promise` or call
`context.done()`, otherwise the following error will be shown:

  [Error]   Error: Choose either to return a promise or call 'done'. Do
  not use both in your script. Learn more:
  https://go.microsoft.com/fwlink/?linkid=2097909

Since we do prefer to return a `Promise`, let's just skip the call to
`context.done()`.

To be sure, these error messages are only shown in the Azure Portal,
when a developer deigns to turn on and look at the logs. Anyway, this
_is_ an error, so let's do fix it.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>